WIP tests fixed
There is also code included to ensure we can run without
any js links, which is a should not happen scenario.

Need to test this later and see if it can be removed again.

Also, Dual Map tests do not work, but this is because
the dual map feature is also broken.